requirements.txt
-f http://pypi.m.sohu.com/
pika
pyyaml
